*: various refactors to increase pgwire and cli reusability
An upcoming balancer process needs to do some small amount of pgwire
decoding and encoding. We would like to reuse existing code instead of
blind copying, because that may make future maintainership harder. To do
this, add a new pgwire-common crate that does not depend on pgrepr or
adapter so it can be used by anything needing pgwire logic. This common
crate allows better abstractions (for example, Severity and Format are
now pgwire types instead of pgrepr and adapter types).

Add a new ore::server crate to abstract out common serving patterns.
This includes moving cli args into common places so they can easily be
reused.

This is all code movement and should contain no logic changes.

See #21843
