+++ src/adapter/Cargo.toml
@@ -20,7 +20,9 @@ futures = "0.3.25"
 hex = "0.4.3"
 itertools = "0.10.5"
 once_cell = "1.16.0"
-launchdarkly-server-sdk = { version = "1.0.0", default_features = false, features = ["hypertls"] }
+launchdarkly-server-sdk = { version = "1.0.0", default_features = false, features = [
+  "hypertls",
+] }
 maplit = "1.0.2"
 mz-audit-log = { path = "../audit-log" }
 mz-build-info = { path = "../build-info" }
@@ -39,6 +41,7 @@ mz-persist-types = { path = "../persist-types" }
 mz-persist-client = { path = "../persist-client" }
 mz-pgcopy = { path = "../pgcopy" }
 mz-pgrepr = { path = "../pgrepr" }
+mz-pgwire-common = { path = "../pgwire-common" }
 mz-postgres-util = { path = "../postgres-util" }
 mz-proto = { path = "../proto" }
 mz-repr = { path = "../repr", features = ["tracing_"] }
@@ -63,7 +66,12 @@ prost = { version = "0.11.3", features = ["no-recursion-limit"] }
 qcell = "0.5"
 rand = "0.8.5"
 rand_chacha = "0.3"
-rdkafka = { version = "0.29.0", features = ["cmake-build", "ssl-vendored", "libz-static", "zstd"] }
+rdkafka = { version = "0.29.0", features = [
+  "cmake-build",
+  "ssl-vendored",
+  "libz-static",
+  "zstd",
+] }
 regex = "1.7.0"
 reqwest = "0.11.13"
 semver = "1.0.16"
@@ -72,7 +80,9 @@ serde_json = "1.0.89"
 serde_plain = "1.0.1"
 smallvec = { version = "1.10.0", features = ["union"] }
 static_assertions = "1.1"
-timely = { version = "0.12.0", default-features = false, features = ["bincode"] }
+timely = { version = "0.12.0", default-features = false, features = [
+  "bincode",
+] }
 tokio = { version = "1.32.0", features = ["rt", "time"] }
 tokio-postgres = { version = "0.7.8" }
 tokio-stream = "0.1.11"
