+++ src/adapter/src/lib.rs
@@ -110,7 +110,6 @@ mod coord;
 mod error;
 mod explain;
 mod notice;
-mod severity;
 mod subscribe;
 mod util;
 
@@ -139,4 +138,3 @@ pub use crate::coord::ExecuteContextExtra;
 pub use crate::coord::{serve, Config};
 pub use crate::error::AdapterError;
 pub use crate::notice::AdapterNotice;
-pub use crate::severity::Severity;
