+++ src/environmentd/src/bin/environmentd/main.rs
@@ -93,16 +93,13 @@ use clap::{ArgEnum, Parser};
 use fail::FailScenario;
 use http::header::HeaderValue;
 use itertools::Itertools;
-use jsonwebtoken::DecodingKey;
 use mz_adapter::catalog::ClusterReplicaSizeMap;
 use mz_aws_secrets_controller::AwsSecretsController;
 use mz_build_info::BuildInfo;
 use mz_cloud_resources::{AwsExternalIdPrefix, CloudResourceController};
 use mz_controller::ControllerConfig;
-use mz_environmentd::{Listeners, ListenersConfig, TlsConfig, BUILD_INFO};
-use mz_frontegg_auth::{
-    Authentication as FronteggAuthentication, AuthenticationConfig as FronteggConfig,
-};
+use mz_environmentd::{Listeners, ListenersConfig, BUILD_INFO};
+use mz_frontegg_auth::FronteggCliArgs;
 use mz_orchestrator::Orchestrator;
 use mz_orchestrator_kubernetes::{
     KubernetesImagePullPolicy, KubernetesOrchestrator, KubernetesOrchestratorConfig,
@@ -115,6 +112,7 @@ use mz_ore::cli::{self, CliConfig, KeyValueArg};
 use mz_ore::error::ErrorExt;
 use mz_ore::metrics::MetricsRegistry;
 use mz_ore::now::SYSTEM_TIME;
+use mz_ore::server::TlsCliArgs;
 use mz_ore::task::RuntimeExt;
 use mz_ore::{halt, metric};
 use mz_persist_client::cache::PersistClientCache;
@@ -135,7 +133,6 @@ use prometheus::IntGauge;
 use tracing::{error, info, warn, Instrument};
 use tracing_opentelemetry::OpenTelemetrySpanExt;
 use url::Url;
-use uuid::Uuid;
 
 mod sys;
 
@@ -238,60 +235,10 @@ pub struct Args {
     /// Wildcards in other positions (e.g., "https://*.foo.com" or "https://foo.*.com") have no effect.
     #[structopt(long, env = "CORS_ALLOWED_ORIGIN")]
     cors_allowed_origin: Vec<HeaderValue>,
-    /// How stringently to demand TLS authentication and encryption.
-    ///
-    /// If set to "disable", then environmentd rejects HTTP and PostgreSQL
-    /// connections that negotiate TLS.
-    ///
-    /// If set to "require", then environmentd requires that all HTTP and
-    /// PostgreSQL connections negotiate TLS. Unencrypted connections will be
-    /// rejected.
-    #[clap(
-        long, env = "TLS_MODE",
-        possible_values = &["disable", "require"],
-        default_value = "disable",
-        default_value_ifs = &[
-            ("frontegg-tenant", None, Some("require")),
-        ],
-        value_name = "MODE",
-    )]
-    tls_mode: String,
-    /// Certificate file for TLS connections.
-    #[clap(
-        long,
-        env = "TLS_CERT",
-        requires = "tls-key",
-        required_if_eq_any(&[("tls-mode", "require")]),
-        value_name = "PATH"
-    )]
-    tls_cert: Option<PathBuf>,
-    /// Private key file for TLS connections.
-    #[clap(
-        long,
-        env = "TLS_KEY",
-        requires = "tls-cert",
-        required_if_eq_any(&[("tls-mode", "require")]),
-        value_name = "PATH"
-    )]
-    tls_key: Option<PathBuf>,
-    /// Enables Frontegg authentication for the specified tenant ID.
-    #[clap(
-        long,
-        env = "FRONTEGG_TENANT",
-        requires_all = &["frontegg-jwk", "frontegg-api-token-url", "frontegg-admin-role"],
-        value_name = "UUID",
-    )]
-    frontegg_tenant: Option<Uuid>,
-    /// JWK used to validate JWTs during Frontegg authentication as a PEM public
-    /// key. Can optionally be base64 encoded with the URL-safe alphabet.
-    #[clap(long, env = "FRONTEGG_JWK", requires = "frontegg-tenant")]
-    frontegg_jwk: Option<String>,
-    /// The full URL (including path) to the Frontegg api-token endpoint.
-    #[clap(long, env = "FRONTEGG_API_TOKEN_URL", requires = "frontegg-tenant")]
-    frontegg_api_token_url: Option<String>,
-    /// The name of the admin role in Frontegg.
-    #[clap(long, env = "FRONTEGG_ADMIN_ROLE", requires = "frontegg-tenant")]
-    frontegg_admin_role: Option<String>,
+    #[clap(flatten)]
+    tls: TlsCliArgs,
+    #[clap(flatten)]
+    frontegg: FronteggCliArgs,
 
     // === Orchestrator options. ===
     /// The service orchestrator implementation to use.
@@ -696,41 +643,8 @@ fn run(mut args: Args) -> Result<(), anyhow::Error> {
     let _failpoint_scenario = FailScenario::setup();
 
     // Configure connections.
-    let tls = if args.tls_mode == "disable" {
-        if args.tls_cert.is_some() {
-            bail!("cannot specify --tls-mode=disable and --tls-cert simultaneously");
-        }
-        if args.tls_key.is_some() {
-            bail!("cannot specify --tls-mode=disable and --tls-key simultaneously");
-        }
-        None
-    } else {
-        let cert = args.tls_cert.unwrap();
-        let key = args.tls_key.unwrap();
-        Some(TlsConfig { cert, key })
-    };
-    let frontegg = match (
-        args.frontegg_tenant,
-        args.frontegg_api_token_url,
-        args.frontegg_jwk,
-        args.frontegg_admin_role,
-    ) {
-        (None, None, None, None) => None,
-        (Some(tenant_id), Some(admin_api_token_url), Some(jwk), Some(admin_role)) => {
-            Some(FronteggAuthentication::new(
-                FronteggConfig {
-                    admin_api_token_url,
-                    decoding_key: DecodingKey::from_rsa_pem(jwk.as_bytes())?,
-                    tenant_id,
-                    now: mz_ore::now::SYSTEM_TIME.clone(),
-                    refresh_before_secs: 60,
-                    admin_role,
-                },
-                mz_frontegg_auth::Client::environmentd_default(),
-            ))
-        }
-        _ => unreachable!("clap enforced"),
-    };
+    let tls = args.tls.into_config()?;
+    let frontegg = args.frontegg.into_auth()?;
 
     // Configure CORS.
     let allowed_origins = if !args.cors_allowed_origin.is_empty() {
