+++ src/frontegg-auth/Cargo.toml
@@ -9,6 +9,7 @@ publish = false
 [dependencies]
 anyhow = "1.0.66"
 base64 = "0.13.1"
+clap = { version = "3.2.24", features = ["wrap_help", "env", "derive"] }
 derivative = "2.2.0"
 jsonwebtoken = "8.2.0"
 mz-ore = { path = "../ore", features = ["network"] }
