+++ src/ore/Cargo.toml
@@ -17,7 +17,9 @@ anyhow = { version = "1.0.66", optional = true }
 # dependencies and is widely considered to be basically part of the stdlib.
 async-trait = { version = "0.1.68", optional = true }
 bytes = { version = "1.3.0", optional = true }
-chrono = { version = "0.4.23", default-features = false, features = ["std"], optional = true }
+chrono = { version = "0.4.23", default-features = false, features = [
+  "std",
+], optional = true }
 clap = { version = "3.2.24", features = ["env"], optional = true }
 ctor = { version = "0.1.26", optional = true }
 either = "1.8.0"
@@ -33,17 +35,29 @@ smallvec = { version = "1.10.0", optional = true }
 stacker = { version = "0.1.15", optional = true }
 sentry = { version = "0.29.1", optional = true, features = ["debug-images"] }
 serde = { version = "1.0.152", features = ["derive"], optional = true }
-tokio = { version = "1.32.0", features = ["io-util", "net", "rt-multi-thread", "sync", "time"], optional = true }
+socket2 = { version = "0.4.7", optional = true }
+tokio = { version = "1.32.0", features = [
+  "io-util",
+  "net",
+  "rt-multi-thread",
+  "sync",
+  "time",
+], optional = true }
 tokio-openssl = { version = "0.6.3", optional = true }
+tokio-stream = { version = "0.1.11", features = ["net"], optional = true }
 # TODO(guswynn): determine, when, if ever, we can remove `tracing-log`
 # The `tracing-log` feature here is load-bearing: While our busiest-logging dependency (`rdkafka`) is now hooked-up
 # to use `tracing`, we cannot remove this feature until we guarantee no dependencies log using the `log` crate, for
 # log lines we care about.
 # Note that this feature is distinct from `tracing`'s `log` feature, which has `tracing` macros emit `log` records if
 # there is no global `tracing` subscriber.
-tracing-subscriber = { version = "0.3.16", default-features = false, features = ["env-filter", "fmt", "json", "tracing-log"], optional = true }
+tracing-subscriber = { version = "0.3.16", default-features = false, features = [
+  "env-filter",
+  "fmt",
+  "json",
+  "tracing-log",
+], optional = true }
 workspace-hack = { version = "0.0.0", path = "../workspace-hack", optional = true }
-
 
 # For the `tracing` feature
 atty = { version = "0.2.14", optional = true }
@@ -55,7 +69,10 @@ tokio-native-tls = { version = "0.3.0", optional = true }
 native-tls = { version = "0.2.11", features = ["alpn"], optional = true }
 hyper = { version = "0.14.23", features = ["http1", "server"], optional = true }
 hyper-tls = { version = "0.5.0", optional = true }
-opentelemetry = { version = "0.20.0", features = ["rt-tokio", "trace"], optional = true }
+opentelemetry = { version = "0.20.0", features = [
+  "rt-tokio",
+  "trace",
+], optional = true }
 opentelemetry-otlp = { version = "0.13.0", optional = true }
 console-subscriber = { version = "0.1.10", optional = true }
 sentry-tracing = { version = "0.29.1", optional = true }
@@ -71,9 +88,17 @@ tokio-test = "0.4.2"
 
 [features]
 default = ["workspace-hack"]
-async = ["async-trait", "futures", "openssl", "tokio-openssl", "tokio", "tracing"]
+async = [
+  "async-trait",
+  "futures",
+  "openssl",
+  "tokio-openssl",
+  "tokio",
+  "tracing",
+]
 bytes_ = ["bytes", "smallvec", "smallvec/const_generics"]
 network = ["async", "bytes", "hyper", "smallvec", "tonic", "tracing"]
+server = ["openssl", "socket2", "tokio-stream"]
 tracing_ = [
   "anyhow",
   "atty",
@@ -87,6 +112,7 @@ tracing_ = [
   "http",
   "hyper",
   "hyper-tls",
+  "metrics",
   "opentelemetry",
   "opentelemetry-otlp",
   "tonic",
