+++ src/pgrepr/Cargo.toml
@@ -14,8 +14,12 @@ dec = "0.4.8"
 once_cell = "1.16.0"
 mz-ore = { path = "../ore" }
 mz-pgrepr-consts = { path = "../pgrepr-consts" }
+mz-pgwire-common = { path = "../pgwire-common" }
 mz-repr = { path = "../repr" }
-postgres-types = { version = "0.2.5", features = ["with-chrono-0_4", "with-uuid-1"] }
+postgres-types = { version = "0.2.5", features = [
+  "with-chrono-0_4",
+  "with-uuid-1",
+] }
 uuid = "1.2.2"
 workspace-hack = { version = "0.0.0", path = "../workspace-hack" }
