+++ src/pgwire-common/Cargo.toml
@@ -1,0 +1,23 @@
+
+[package]
+name = "mz-pgwire-common"
+description = "Common logic for the PostgreSQL wire protocol."
+version = "0.0.0"
+edition.workspace = true
+rust-version.workspace = true
+publish = false
+
+[dependencies]
+async-trait = "0.1.68"
+byteorder = "1.4.3"
+bytes = "1.3.0"
+bytesize = "1.1.0"
+mz-ore = { path = "../ore", features = ["network"] }
+tokio = "1.24.2"
+tokio-openssl = "0.6.3"
+tokio-postgres = { version = "0.7.8" }
+tracing = "0.1.37"
+workspace-hack = { version = "0.0.0", path = "../workspace-hack" }
+
+[package.metadata.cargo-udeps.ignore]
+normal = ["workspace-hack"]
