+++ src/pgwire-common/src/conn.rs
@@ -1,0 +1,77 @@
+// Copyright Materialize, Inc. and contributors. All rights reserved.
+//
+// Use of this software is governed by the Business Source License
+// included in the LICENSE file.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0.
+
+use std::pin::Pin;
+use std::task::{Context, Poll};
+
+use async_trait::async_trait;
+use mz_ore::netio::AsyncReady;
+use tokio::io::{self, AsyncRead, AsyncWrite, Interest, ReadBuf, Ready};
+use tokio_openssl::SslStream;
+
+#[derive(Debug)]
+pub enum Conn<A> {
+    Unencrypted(A),
+    Ssl(SslStream<A>),
+}
+
+impl<A> AsyncRead for Conn<A>
+where
+    A: AsyncRead + AsyncWrite + Unpin,
+{
+    fn poll_read(
+        self: Pin<&mut Self>,
+        cx: &mut Context,
+        buf: &mut ReadBuf,
+    ) -> Poll<io::Result<()>> {
+        match self.get_mut() {
+            Conn::Unencrypted(inner) => Pin::new(inner).poll_read(cx, buf),
+            Conn::Ssl(inner) => Pin::new(inner).poll_read(cx, buf),
+        }
+    }
+}
+
+impl<A> AsyncWrite for Conn<A>
+where
+    A: AsyncRead + AsyncWrite + Unpin,
+{
+    fn poll_write(self: Pin<&mut Self>, cx: &mut Context, buf: &[u8]) -> Poll<io::Result<usize>> {
+        match self.get_mut() {
+            Conn::Unencrypted(inner) => Pin::new(inner).poll_write(cx, buf),
+            Conn::Ssl(inner) => Pin::new(inner).poll_write(cx, buf),
+        }
+    }
+
+    fn poll_flush(self: Pin<&mut Self>, cx: &mut Context) -> Poll<io::Result<()>> {
+        match self.get_mut() {
+            Conn::Unencrypted(inner) => Pin::new(inner).poll_flush(cx),
+            Conn::Ssl(inner) => Pin::new(inner).poll_flush(cx),
+        }
+    }
+
+    fn poll_shutdown(self: Pin<&mut Self>, cx: &mut Context) -> Poll<io::Result<()>> {
+        match self.get_mut() {
+            Conn::Unencrypted(inner) => Pin::new(inner).poll_shutdown(cx),
+            Conn::Ssl(inner) => Pin::new(inner).poll_shutdown(cx),
+        }
+    }
+}
+
+#[async_trait]
+impl<A> AsyncReady for Conn<A>
+where
+    A: AsyncRead + AsyncWrite + AsyncReady + Sync + Unpin,
+{
+    async fn ready(&self, interest: Interest) -> io::Result<Ready> {
+        match self {
+            Conn::Unencrypted(inner) => inner.ready(interest).await,
+            Conn::Ssl(inner) => inner.ready(interest).await,
+        }
+    }
+}
