+++ src/pgwire-common/src/message.rs
@@ -1,0 +1,259 @@
+// Copyright Materialize, Inc. and contributors. All rights reserved.
+//
+// Use of this software is governed by the Business Source License
+// included in the LICENSE file.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0.
+
+use std::collections::BTreeMap;
+
+use tokio_postgres::error::SqlState;
+
+use crate::{Format, Severity};
+
+// Pgwire protocol versions are represented as 32-bit integers, where the
+// high 16 bits represent the major version and the low 16 bits represent the
+// minor version.
+//
+// There have only been three released protocol versions, v1.0, v2.0, and v3.0.
+// The protocol changes very infrequently: the most recent protocol version,
+// v3.0, was released with Postgres v7.4 in 2003.
+//
+// Somewhat unfortunately, the protocol overloads the version field to indicate
+// special types of connections, namely, SSL connections and cancellation
+// connections. These pseudo-versions were constructed to avoid ever matching
+// a true protocol version.
+
+pub const VERSION_1: i32 = 0x10000;
+pub const VERSION_2: i32 = 0x20000;
+pub const VERSION_3: i32 = 0x30000;
+pub const VERSION_CANCEL: i32 = (1234 << 16) + 5678;
+pub const VERSION_SSL: i32 = (1234 << 16) + 5679;
+pub const VERSION_GSSENC: i32 = (1234 << 16) + 5680;
+
+pub const VERSIONS: &[i32] = &[
+    VERSION_1,
+    VERSION_2,
+    VERSION_3,
+    VERSION_CANCEL,
+    VERSION_SSL,
+    VERSION_GSSENC,
+];
+
+/// Like [`FrontendMessage`], but only the messages that can occur during
+/// startup protocol negotiation.
+#[derive(Debug)]
+pub enum FrontendStartupMessage {
+    /// Begin a connection.
+    Startup {
+        version: i32,
+        params: BTreeMap<String, String>,
+    },
+
+    /// Request SSL encryption for the connection.
+    SslRequest,
+
+    /// Request GSSAPI encryption for the connection.
+    GssEncRequest,
+
+    /// Cancel a query that is running on another connection.
+    CancelRequest {
+        /// The target connection ID.
+        conn_id: u32,
+        /// The secret key for the target connection.
+        secret_key: u32,
+    },
+}
+
+/// A decoded frontend pgwire [message], representing instructions for the
+/// backend.
+///
+/// [message]: https://www.postgresql.org/docs/11/protocol-message-formats.html
+#[derive(Debug)]
+pub enum FrontendMessage {
+    /// Execute the specified SQL.
+    ///
+    /// This is issued as part of the simple query flow.
+    Query {
+        /// The SQL to execute.
+        sql: String,
+    },
+
+    /// Parse the specified SQL into a prepared statement.
+    ///
+    /// This starts the extended query flow.
+    Parse {
+        /// The name of the prepared statement to create. An empty string
+        /// specifies the unnamed prepared statement.
+        name: String,
+        /// The SQL to parse.
+        sql: String,
+        /// The OID of each parameter data type for which the client wants to
+        /// prespecify types. A zero OID is equivalent to leaving the type
+        /// unspecified.
+        ///
+        /// The number of specified parameter data types can be less than the
+        /// number of parameters specified in the query.
+        param_types: Vec<u32>,
+    },
+
+    /// Describe an existing prepared statement.
+    ///
+    /// This command is part of the extended query flow.
+    DescribeStatement {
+        /// The name of the prepared statement to describe.
+        name: String,
+    },
+
+    /// Describe an existing portal.
+    ///
+    /// This command is part of the extended query flow.
+    DescribePortal {
+        /// The name of the portal to describe.
+        name: String,
+    },
+
+    /// Bind an existing prepared statement to a portal.
+    ///
+    /// This command is part of the extended query flow.
+    Bind {
+        /// The destination portal. An empty string selects the unnamed
+        /// portal. The portal can later be executed with the `Execute` command.
+        portal_name: String,
+        /// The source prepared statement. An empty string selects the unnamed
+        /// prepared statement.
+        statement_name: String,
+        /// The formats used to encode the parameters in `raw_parameters`.
+        param_formats: Vec<Format>,
+        /// The value of each parameter, encoded using the formats described
+        /// by `parameter_formats`.
+        raw_params: Vec<Option<Vec<u8>>>,
+        /// The desired formats for the columns in the result set.
+        result_formats: Vec<Format>,
+    },
+
+    /// Execute a bound portal.
+    ///
+    /// This command is part of the extended query flow.
+    Execute {
+        /// The name of the portal to execute.
+        portal_name: String,
+        /// The maximum number number of rows to return before suspending.
+        ///
+        /// 0 or negative means infinite.
+        max_rows: i32,
+    },
+
+    /// Flush any pending output.
+    ///
+    /// This command is part of the extended query flow.
+    Flush,
+
+    /// Finish an extended query.
+    ///
+    /// This command is part of the extended query flow.
+    Sync,
+
+    /// Close the named statement.
+    ///
+    /// This command is part of the extended query flow.
+    CloseStatement {
+        name: String,
+    },
+
+    /// Close the named portal.
+    ///
+    // This command is part of the extended query flow.
+    ClosePortal {
+        name: String,
+    },
+
+    /// Terminate a connection.
+    Terminate,
+
+    CopyData(Vec<u8>),
+
+    CopyDone,
+
+    CopyFail(String),
+
+    Password {
+        password: String,
+    },
+}
+
+impl FrontendMessage {
+    pub fn name(&self) -> &'static str {
+        match self {
+            FrontendMessage::Query { .. } => "query",
+            FrontendMessage::Parse { .. } => "parse",
+            FrontendMessage::DescribeStatement { .. } => "describe_statement",
+            FrontendMessage::DescribePortal { .. } => "describe_portal",
+            FrontendMessage::Bind { .. } => "bind",
+            FrontendMessage::Execute { .. } => "execute",
+            FrontendMessage::Flush => "flush",
+            FrontendMessage::Sync => "sync",
+            FrontendMessage::CloseStatement { .. } => "close_statement",
+            FrontendMessage::ClosePortal { .. } => "close_portal",
+            FrontendMessage::Terminate => "terminate",
+            FrontendMessage::CopyData(_) => "copy_data",
+            FrontendMessage::CopyDone => "copy_done",
+            FrontendMessage::CopyFail(_) => "copy_fail",
+            FrontendMessage::Password { .. } => "password",
+        }
+    }
+}
+
+#[derive(Debug)]
+pub struct ErrorResponse {
+    pub severity: Severity,
+    pub code: SqlState,
+    pub message: String,
+    pub detail: Option<String>,
+    pub hint: Option<String>,
+    pub position: Option<usize>,
+}
+
+impl ErrorResponse {
+    pub fn fatal<S>(code: SqlState, message: S) -> ErrorResponse
+    where
+        S: Into<String>,
+    {
+        ErrorResponse::new(Severity::Fatal, code, message)
+    }
+
+    pub fn error<S>(code: SqlState, message: S) -> ErrorResponse
+    where
+        S: Into<String>,
+    {
+        ErrorResponse::new(Severity::Error, code, message)
+    }
+
+    pub fn notice<S>(code: SqlState, message: S) -> ErrorResponse
+    where
+        S: Into<String>,
+    {
+        ErrorResponse::new(Severity::Notice, code, message)
+    }
+
+    fn new<S>(severity: Severity, code: SqlState, message: S) -> ErrorResponse
+    where
+        S: Into<String>,
+    {
+        ErrorResponse {
+            severity,
+            code,
+            message: message.into(),
+            detail: None,
+            hint: None,
+            position: None,
+        }
+    }
+
+    pub fn with_position(mut self, position: usize) -> ErrorResponse {
+        self.position = Some(position);
+        self
+    }
+}
