+++ src/pgwire/src/server.rs
@@ -11,19 +11,22 @@ use std::fmt;
 use std::future::Future;
 use std::pin::Pin;
 use std::sync::{Arc, Mutex};
-use std::task::{Context, Poll};
 
 use async_trait::async_trait;
 use mz_frontegg_auth::Authentication as FronteggAuthentication;
 use mz_ore::netio::AsyncReady;
+use mz_ore::server::{ConnectionHandler, TlsConfig};
+use mz_pgwire_common::{
+    decode_startup, Conn, FrontendStartupMessage, ACCEPT_SSL_ENCRYPTION, REJECT_ENCRYPTION,
+};
 use mz_sql::session::vars::ConnectionCounter;
-use openssl::ssl::{Ssl, SslContext};
-use tokio::io::{self, AsyncRead, AsyncWrite, AsyncWriteExt, Interest, ReadBuf, Ready};
+use openssl::ssl::Ssl;
+use tokio::io::{AsyncRead, AsyncWrite, AsyncWriteExt};
+use tokio::net::TcpStream;
 use tokio_openssl::SslStream;
 use tracing::trace;
 
-use crate::codec::{self, FramedConn, ACCEPT_SSL_ENCRYPTION, REJECT_ENCRYPTION};
-use crate::message::FrontendStartupMessage;
+use crate::codec::FramedConn;
 use crate::metrics::{Metrics, MetricsConfig};
 use crate::protocol;
 
@@ -52,24 +55,6 @@ pub struct Config {
     pub active_connection_count: Arc<Mutex<ConnectionCounter>>,
 }
 
-/// Configures a server's TLS encryption and authentication.
-#[derive(Clone, Debug)]
-pub struct TlsConfig {
-    /// The SSL context used to manage incoming TLS negotiations.
-    pub context: SslContext,
-    /// The TLS mode.
-    pub mode: TlsMode,
-}
-
-/// Specifies how strictly to enforce TLS encryption.
-#[derive(Debug, Clone, Copy)]
-pub enum TlsMode {
-    /// Allow TLS encryption.
-    Allow,
-    /// Require that clients negotiate TLS encryption.
-    Require,
-}
-
 /// A server that communicates with clients via the pgwire protocol.
 pub struct Server {
     tls: Option<TlsConfig>,
@@ -78,6 +63,18 @@ pub struct Server {
     metrics: Metrics,
     internal: bool,
     active_connection_count: Arc<Mutex<ConnectionCounter>>,
+}
+
+#[async_trait]
+impl mz_ore::server::Server for Server {
+    const NAME: &'static str = "pgwire";
+
+    fn handle_connection(&self, conn: TcpStream) -> ConnectionHandler {
+        // Using fully-qualified syntax means we won't accidentally call
+        // ourselves (i.e., silently infinitely recurse) if the name or type of
+        // `crate::Server::handle_connection` changes.
+        Box::pin(crate::Server::handle_connection(self, conn))
+    }
 }
 
 impl Server {
@@ -115,7 +112,7 @@ impl Server {
                     let conn_id = adapter_client.new_conn_id()?;
                     let mut conn = Conn::Unencrypted(conn);
                     loop {
-                        let message = codec::decode_startup(&mut conn).await?;
+                        let message = decode_startup(&mut conn).await?;
 
                         match &message {
                             Some(message) => trace!("cid={} recv={:?}", conn_id, message),
@@ -186,67 +183,6 @@ impl Server {
             .await;
             metrics.connection_status(result.is_ok()).inc();
             result
-        }
-    }
-}
-
-#[derive(Debug)]
-pub enum Conn<A> {
-    Unencrypted(A),
-    Ssl(SslStream<A>),
-}
-
-impl<A> AsyncRead for Conn<A>
-where
-    A: AsyncRead + AsyncWrite + Unpin,
-{
-    fn poll_read(
-        self: Pin<&mut Self>,
-        cx: &mut Context,
-        buf: &mut ReadBuf,
-    ) -> Poll<io::Result<()>> {
-        match self.get_mut() {
-            Conn::Unencrypted(inner) => Pin::new(inner).poll_read(cx, buf),
-            Conn::Ssl(inner) => Pin::new(inner).poll_read(cx, buf),
-        }
-    }
-}
-
-impl<A> AsyncWrite for Conn<A>
-where
-    A: AsyncRead + AsyncWrite + Unpin,
-{
-    fn poll_write(self: Pin<&mut Self>, cx: &mut Context, buf: &[u8]) -> Poll<io::Result<usize>> {
-        match self.get_mut() {
-            Conn::Unencrypted(inner) => Pin::new(inner).poll_write(cx, buf),
-            Conn::Ssl(inner) => Pin::new(inner).poll_write(cx, buf),
-        }
-    }
-
-    fn poll_flush(self: Pin<&mut Self>, cx: &mut Context) -> Poll<io::Result<()>> {
-        match self.get_mut() {
-            Conn::Unencrypted(inner) => Pin::new(inner).poll_flush(cx),
-            Conn::Ssl(inner) => Pin::new(inner).poll_flush(cx),
-        }
-    }
-
-    fn poll_shutdown(self: Pin<&mut Self>, cx: &mut Context) -> Poll<io::Result<()>> {
-        match self.get_mut() {
-            Conn::Unencrypted(inner) => Pin::new(inner).poll_shutdown(cx),
-            Conn::Ssl(inner) => Pin::new(inner).poll_shutdown(cx),
-        }
-    }
-}
-
-#[async_trait]
-impl<A> AsyncReady for Conn<A>
-where
-    A: AsyncRead + AsyncWrite + AsyncReady + Sync + Unpin,
-{
-    async fn ready(&self, interest: Interest) -> io::Result<Ready> {
-        match self {
-            Conn::Unencrypted(inner) => inner.ready(interest).await,
-            Conn::Ssl(inner) => inner.ready(interest).await,
         }
     }
 }
